src/fault-detection/fault_detection_job/requirements.txt
pandas
numpy
python-dotenv
pymongo
gcsfs
pandas-gbq
google-cloud-bigquery
tensorflow
joblib
scikit-learn

